create new file to account_tax_template
