+++ l10n_br_account_product/models/__init__.py
@@ -11,6 +11,7 @@ from . import l10n_br_account
 from . import account_product_fiscal_classification
 from . import account_invoice
 from . import account_payment_term
+from . import account_tax_template
 from . import account
 from . import product
 from . import res_country
