+++ l10n_br_account_product/models/account.py
@@ -8,23 +8,6 @@ from openerp import models, fields, api
 from openerp.tools import DEFAULT_SERVER_DATE_FORMAT
 
 
-class AccountTaxTemplate(models.Model):
-    """Implement computation method in taxes"""
-    _inherit = 'account.tax.template'
-
-    icms_base_type = fields.Selection(
-        [('0', 'Margem Valor Agregado (%)'), ('1', 'Pauta (valor)'),
-         ('2', 'Preço Tabelado Máximo (valor)'),
-         ('3', 'Valor da Operação')],
-        'Tipo Base ICMS', required=True, default='0')
-    icms_st_base_type = fields.Selection(
-        [('0', 'Preço tabelado ou máximo  sugerido'),
-         ('1', 'Lista Negativa (valor)'),
-         ('2', 'Lista Positiva (valor)'), ('3', 'Lista Neutra (valor)'),
-         ('4', 'Margem Valor Agregado (%)'), ('5', 'Pauta (valor)')],
-        'Tipo Base ICMS ST', required=True, default='4')
-
-
 class AccountTax(models.Model):
     """Implement computation method in taxes"""
     _inherit = 'account.tax'
