+++ l10n_br_account_product/models/account_tax_template.py
@@ -1,0 +1,22 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2013  Renato Lima - Akretion
+# License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
+
+from openerp import models, fields
+
+
+class AccountTaxTemplate(models.Model):
+    """Implement computation method in taxes"""
+    _inherit = 'account.tax.template'
+
+    icms_base_type = fields.Selection(
+        [('0', 'Margem Valor Agregado (%)'), ('1', 'Pauta (valor)'),
+         ('2', 'Preço Tabelado Máximo (valor)'),
+         ('3', 'Valor da Operação')],
+        'Tipo Base ICMS', required=True, default='0')
+    icms_st_base_type = fields.Selection(
+        [('0', 'Preço tabelado ou máximo  sugerido'),
+         ('1', 'Lista Negativa (valor)'),
+         ('2', 'Lista Positiva (valor)'), ('3', 'Lista Neutra (valor)'),
+         ('4', 'Margem Valor Agregado (%)'), ('5', 'Pauta (valor)')],
+        'Tipo Base ICMS ST', required=True, default='4')
